Add -fflags bitexact to youtube-dl 'extract video' args
Modify bs1770gain pass. Previously it calculated the RG info *and*
applied it to the file (using ffmpeg, but it didn't use -fflags bitexact
so the muxing of opus files always resulted in a not-bit-identical
output). Now it calculates the RG info, prints it in XML to stdout. That
output is now parsed and applied with tagit to the file.